5-tier-application+ingres/Go+Python+React+Postgres+Rabbitmq/user-service/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
pika==1.3.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
pika==1.3.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
